firebase deployment and modal fix

+++ client/src/components/nominationList/nominationCard/NominationCard.tsx
@@ -1,4 +1,5 @@
-import React, { useRef } from "react";
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect, useState } from "react";
 import { Button } from "@shopify/polaris";
 import { useDispatch, useSelector } from "react-redux";
 import INomination from "../../../models/interfaces/Nomination";
@@ -9,33 +10,41 @@ import "./NominationCard.css";
 const reduxActions = new ReduxActions();
 
 const NominationCard = (props: INomination): JSX.Element => {
-  const { Title } = props;
+  const { Title, Year } = props;
   const nominationList = useSelector(
     (state: RootState) => state.nominationList,
   );
-  const nominationCardRef = useRef<any>();
-  nominationCardRef.current = "nominationCard-show";
+  const [showNomination, setShowNomination] = useState({ appear: false, delete: false });
   const dispatch = useDispatch();
 
-  const removeNominationFromList = (): void => {
-    nominationCardRef.current = "nominationCard-disappear";
-    setTimeout(() => {
+  useEffect(() => {
+    if (showNomination.delete) {
       dispatch(reduxActions.removeNomination(props));
       window.localStorage.setItem("nominations", JSON.stringify(nominationList));
-      dispatch(reduxActions.changeNominationList());
-    }, 300);
+      setTimeout(() => {
+        dispatch(reduxActions.changeNominationList());
+      }, 200);
+    } else {
+      setTimeout(() => {
+        setShowNomination({ appear: true, delete: false });
+      }, 200);
+    }
+  }, [showNomination]);
+
+  const removeNominationFromList = (): void => {
+    setShowNomination({ appear: false, delete: true });
   };
 
   return (
-    <li className={nominationCardRef.current}>
+    <li className={showNomination.appear ? "nominationCard" : "nominationCard-disappear"}>
       <div className="nominationCard">
-        {Title}
+        {`${Title} (${Year})`}
         <div className="spacing" />
         <Button
           destructive
           onClick={removeNominationFromList}
         >
-          Remove
+          -
         </Button>
       </div>
     </li>
